refactor(uebermensch): extract llm prompts/schemas to lib/llm-prompts.ts
Pure code movement — no behavior change. Splits the 1020-LoC KernelLlm.ts so the next LlmPort adapters (AnthropicLlm, LMStudioLlm) can share prompt strings, output schemas, and JSON post-processing without copy-paste drift.

Moved to lib/llm-prompts.ts:
- System prompts: SYSTEM_PROMPT (brief), SUBTOPIC_SYSTEM_PROMPT, REPORT_SYSTEM_PROMPT, SUMMARY_SYSTEM_PROMPT, RESEARCH_SYSTEM_PROMPT
- User-prompt builders: buildUserPrompt, buildSubtopicUserPrompt, buildInterestReportUserPrompt, buildSummaryUserPrompt, buildResearchQueryUserPrompt
- Effect schemas: LlmOutputSchema, SubtopicProposeOutputSchema, InterestReportOutputSchema (+ ItemSchema, SubtopicProposalSchema as private)
- Response post-processing: extractJson, normalizeInsights, decodeLlmJson
- JSON response_format helpers: JsonResponseFormat, briefJsonFormat, subtopicJsonFormat, interestReportJsonFormat
- Constants: MAX_CANDIDATE_EXCERPT, SUMMARY_INPUT_CHARS_CAP, SUMMARY_MAX_TOKENS

Stayed in adapters/KernelLlm.ts:
- HTTP transport (fetchKernel, postAnthropic, postWorkersAi, postLlm)
- Effort/thinking config (effortConfigFor, thinkingBody)
- Anthropic cost rates + costForResponse
- Model selection (modelFor, summaryModelFor, isAnthropicModel/isWorkersAiModel)
- Identity headers (sessionIdFor, SERVICE_NAME)
- Kernel error mapping (kernelDownErr, classifyKernelStatus, isConnRefused)
- Layer.succeed binding to LlmService

Test stability: kept the `_internal` re-export shape identical so existing tests/kernel-llm.test.ts doesn't change. Tests will be migrated to import directly from lib/llm-prompts.ts in the AnthropicLlm follow-up.

First slice of PR-A (uebermensch ejection plan, three-PR sequence). Resolves reviewer concern C5 (prompt-drift risk) and unblocks AnthropicLlm + LMStudioLlm direct adapters.

220/220 vitest pass; tsc --noEmit clean.
